gonna try one webgl canvas with many views
