+++ src/bootstrap/index.html
@@ -112,6 +112,7 @@ input::placeholder {
 
     <div id="canvas-container">
       <div id="plugins"></div>
+      <div id="webgl"></div>
       <div id="windows"></div>
     </div>
     <div id="statusline"></div>
